+++ backends/stablehlo/builder.go
@@ -1,0 +1,34 @@
+package stablehlo
+
+import (
+	"github.com/gomlx/gomlx/backends"
+	"github.com/gomlx/gomlx/backends/notimplemented"
+	"github.com/gomlx/stablehlo"
+	"k8s.io/klog/v2"
+)
+
+// Builder keeps track of the computation graph being defined.
+type Builder struct {
+	notimplemented.Builder
+
+	name     string
+	backend  *Backend
+	compiled bool
+
+	builder *stablehlo.Builder
+}
+
+var _ backends.Builder = (*Builder)(nil)
+
+// Builder creates a new builder used to define a new computation.
+func (backend *Backend) Builder(name string) backends.Builder {
+	if err := backend.CheckValid(); err != nil {
+		klog.Error(err)
+		return nil
+	}
+	return &Builder{
+		backend: backend,
+		builder: stablehlo.New(name),
+		name:    name,
+	}
+}